게임 프레임 구현 및 떨어지는 속도 조절 + 속도 short -> float 변경

+++ Project1/Block.cpp
@@ -59,7 +59,7 @@ void Block::SetRandomBlock()	//블럭 랜덤 고르기
 	}
 }
 
-void Block::SetBlockXY(int x, int y)	//위치변화
+void Block::SetBlockXY(float x, float y)	//위치변화
 {
 	Y = Y + y;
 	X = X + x;
